requirements.txt
aniso8601==10.0.0
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
flasgger==0.9.7.1
Flask==3.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
flask-swagger-ui==4.11.1
googleapis-common-protos==1.66.0
grpcio==1.70.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.1
opentelemetry-api==1.30.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-exporter-prometheus==0.51b0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.43b0
packaging==24.2
prometheus_client==0.21.1
protobuf==5.29.3
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
